Add tournament delete API and runtime functions.

+++ console/ui/src/app/leaderboard/leaderboard.component.ts
@@ -50,12 +50,16 @@ export class LeaderboardComponent implements OnInit {
   deleteLeaderboard(event): void {
     event.target.disabled = true;
     this.error = '';
-    this.consoleService.deleteLeaderboard('', this.leaderboard.id).subscribe(() => {
-      this.error = '';
-      this.router.navigate(['/leaderboards']);
-    }, err => {
-      this.error = err;
-    });
+    if (this.leaderboard.tournament) {
+     // TODO: Call tournament delete
+    } else {
+      this.consoleService.deleteLeaderboard('', this.leaderboard.id).subscribe(() => {
+        this.error = '';
+        this.router.navigate(['/leaderboards']);
+      }, err => {
+        this.error = err;
+      });
+    }
   }
 
   deleteAllowed(): boolean {
